refactor(hangar): move the transcript sink into its own module for privacy
Rust privacy is per-module, so a StoreWriter wrapped inside acp_pool.rs
was still reachable from every SessionActor method: the invariant was a
sign on the door, not a lock. Proved it by probe, self.sink.writer
compiled.

The sink now lives in acp_transcript.rs, where the writer field is
unreachable from acp_pool entirely. The same probe is now E0616. A new
flush that skipped the live publish cannot compile.

That is what the module is for; it is not a size refactor.
