+++ ainb-tui/crates/ainb-hangar-daemon/src/acp_pool.rs
@@ -99,9 +99,7 @@ use ainb_acp::reducer::TranscriptReducer;
 use ainb_acp::store_writer::{HighWater, Lifecycle, StoreWriter, WriterConfig};
 use ainb_hangar_core::clock::{HangarClock, SystemClock};
 use ainb_hangar_core::idgen::{IdGen, SystemIdGen};
-use ainb_hangar_proto::events::MessageKind;
 use ainb_hangar_proto::settings::{AcpPoolHealth, AcpProcessHealth, AcpSessionHealth};
-use ainb_hangar_proto::transcript::AcpClassifier;
 use ainb_hangar_store::Store;
 use ainb_hangar_store::repo::attention::AttentionRepo;
 use ainb_hangar_store::repo::fleet::{
@@ -112,11 +110,12 @@ use ainb_hangar_store::repo::fleet_acp_session::{
 };
 use ainb_hangar_store::repo::fleet_message::{FleetMessageRepo, NewFleetMessage};
 use ainb_hangar_store::repo::fleet_provider_event::{
-    FleetProviderEventError, FleetProviderEventRepo, NewFleetProviderEvent,
+    FleetProviderEventRepo, NewFleetProviderEvent,
 };
 use sqlx::SqlitePool;
 use tokio::sync::{Semaphore, mpsc, oneshot};
 
+use crate::acp_transcript::TranscriptSink;
 use crate::events::EventSink;
 use crate::runner::RunStream;
 use tracing::Instrument as _;
@@ -1801,97 +1800,6 @@ struct SessionActor {
 
 /// What one turn's `session/prompt` came back with.
 type TurnResult = Result<PromptResponse, AcpError>;
-
-/// One session's transcript, live and durable, behind one door.
-///
-/// The [`StoreWriter`] is PRIVATE to this type on purpose. Every durable
-/// transcript row must also be published live, or the durable re-read carries
-/// lines the operator never saw stream — and that omission has now been written
-/// twice, both times by someone adding a perfectly legitimate new flush (the
-/// turn-end one, then the adapter-death drain). A doc comment did not stop the
-/// second. Here the only ways in are [`Self::chunk`] and [`Self::lifecycle`],
-/// both of which publish before they write, so a third flush cannot compile
-/// without going through one of them.
-///
-/// Live is published BEFORE the durable commit, deliberately: the writer
-/// buffers and commits on a cadence, so publishing after would hold the
-/// operator's view back by up to a flush interval. See the module docs for the
-/// one guarantee that costs.
-struct TranscriptSink {
-    writer: StoreWriter,
-    /// Where this session's rows go live, for a TASK session only. `None` for
-    /// every chat session: no task to name, and its own stream elsewhere.
-    stream: Option<RunStream>,
-    /// The SAME classifier the durable `board_card_timeline` read uses, so a
-    /// line published live and its re-read twin are byte-identical.
-    classifier: AcpClassifier,
-    /// Tool calls published this turn, for the run banner's tally.
-    tool_calls: u32,
-}
-
-impl TranscriptSink {
-    fn new(writer: StoreWriter) -> Self {
-        Self {
-            writer,
-            stream: None,
-            classifier: AcpClassifier::default(),
-            tool_calls: 0,
-        }
-    }
-
-    /// Publish a chunk live, then commit it.
-    async fn chunk(
-        &mut self,
-        chunk: &ainb_acp::reducer::TranscriptChunk,
-    ) -> Result<Option<HighWater>, FleetProviderEventError> {
-        self.publish(chunk.kind.event_type(), &chunk.payload);
-        self.writer.push(chunk).await
-    }
-
-    /// Publish a lifecycle marker live, then commit it.
-    async fn lifecycle(
-        &mut self,
-        marker: Lifecycle,
-        payload: serde_json::Value,
-    ) -> Result<Option<HighWater>, FleetProviderEventError> {
-        self.publish(marker.event_type(), &payload);
-        self.writer.lifecycle(marker, payload).await
-    }
-
-    /// Classify one row and stream every line it yields.
-    fn publish(&mut self, event_type: &str, payload: &serde_json::Value) {
-        if self.stream.is_none() {
-            return;
-        }
-        for (kind, body) in self.classifier.classify_value(event_type, payload) {
-            if kind == MessageKind::ToolCall {
-                self.tool_calls = self.tool_calls.saturating_add(1);
-            }
-            if let Some(stream) = &self.stream {
-                stream.line(kind, body);
-            }
-        }
-    }
-
-    /// The run banner's tally + clock. No-op for a chat session.
-    fn progress(&self, elapsed: Duration) {
-        if let Some(stream) = &self.stream {
-            stream.progress(self.tool_calls, elapsed);
-        }
-    }
-
-    async fn tick(&mut self) -> Result<Option<HighWater>, FleetProviderEventError> {
-        self.writer.tick().await
-    }
-
-    fn bytes_written(&self) -> u64 {
-        self.writer.bytes_written()
-    }
-
-    fn set_acp_session_id(&mut self, id: Option<String>) {
-        self.writer.set_acp_session_id(id);
-    }
-}
 
 /// Bind a session's live task stream, or `None` when the scope is not a task's.
 ///
